+++ CHANGELOG.md
@@ -12,9 +12,10 @@ and this project adheres to [Semantic Versioning][].
 
 ### Fixed
 
--   Fix color assignment for NaN values (#257)
--   Fix channel str support (#221)
+-   Fixed channel str support (#221)
+-   Fixed color assignment for NaN values (#257)
 -   Updated incorrect link to documentation (#261)
+-   Fixed plotting of categorical data (#262) 
 
 ## [0.2.2] - 2024-05-02